fix: audit MEDIUM + LOW polish — A-4, F-11, F-12, F-14, F-16, C-4
A-4 (LOW): bind URL's classId to URL's competitionId on
/competitions/[competitionId]/classes/[classId]/{entries,results}
GET + entries POST + entries/[entryId] PATCH so a stale
competitions/<X>/classes/<Y>/... link returns 404 if X is not
Y's parent (matching the existing class-edit and results-POST
binding).

F-11 (MED): replaced 7 inline staleTime literals (5 * 60 * 1000,
60 * 1000, 30_000, 10_000) with shared constants —
STALE_TIME_STABLE / STALE_TIME_MEDIUM / STALE_TIME_FREQUENT /
STALE_TIME_BURST. Added STALE_TIME_MEDIUM (1m) and
STALE_TIME_BURST (10s) to packages/shared/src/constants.

F-12 (LOW): added MS_PER_DAY / MS_PER_YEAR_AVG /
ACTIVE_CLUB_COOKIE_TTL_SECONDS to shared constants and replaced
the duplicated `60 * 60 * 24 * 30` cookie-TTL literals in
tenant.ts and /me/active-club/route.ts.

F-14 (LOW): paginationSchema now uses DEFAULT_PAGE_SIZE +
MAX_PAGE_SIZE constants instead of hardcoded 25/100.

F-15 (LOW): added RATE_LIMIT_TIGHT / STANDARD / PUBLIC_BURST /
HEALTHCHECK presets to shared constants. Existing call-sites
keep their inline objects; new routes can adopt the presets.

F-16 (LOW): audit_log count query now casts to ::int, mirroring
every other count query in the codebase. The `<number>` generic
was lying — Postgres returns count(*) as bigint which Neon
serialises as a string.

C-4 (LOW): replaced unprefixed `'crypto'` imports with
`'node:crypto'` in apps/web/app/api/v1/{owners,staff}/route.ts
and packages/db/src/queries/riders.ts to match the rest of the
codebase. Workers nodejs_compat covers both, but the prefixed
form is what Cloudflare's docs recommend and what the rest of
the lib uses.

+++ packages/db/src/queries/audit-log.ts
@@ -82,7 +82,11 @@ export async function getAuditLog(clubId: string, filters: AuditLogFilters) {
       .limit(filters.pageSize)
       .offset(offset),
     db
-      .select({ count: sql<number>`count(*)` })
+      // ::int cast — Postgres returns count(*) as bigint which @neondatabase/
+      // serverless surfaces as a string in the JSON output; the `<number>`
+      // generic was lying. Mirrors every other count query in the codebase.
+      // See audit F-16.
+      .select({ count: sql<number>`count(*)::int` })
       .from(auditLog)
       .where(and(...conditions)),
   ]);
